requirements.txt
# ── Core Framework ──────────────────────────────────────────────
streamlit>=1.33.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0

# ── AI / LLM ───────────────────────────────────────────────────
openai>=1.12.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
faiss-cpu>=1.10.0
tiktoken>=0.6.0

# ── Voice ───────────────────────────────────────────────────────
gtts>=2.5.1
pydub>=0.25.1

# ── Database ────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0

# ── Images / Vision ────────────────────────────────────────────
Pillow>=10.0.0

# ── WebSockets ──────────────────────────────────────────────────
websockets>=12.0

# ── Auth / Security ────────────────────────────────────────────
pyjwt>=2.8.0
# Password hashing uses stdlib hashlib (PBKDF2-SHA256) — no external dep needed

# ── HTTP Client ─────────────────────────────────────────────────
httpx>=0.27.0

# ── Logging / Monitoring ───────────────────────────────────────
structlog>=24.1.0

# ── Testing ─────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
